MS-1454 task: update naming conventions + comment descriptions
- rename files for better readability
- rename methods for better readability
- rename test name and route name
- update method comments to more verbose
